CORE-4354: Refactor V1 packaging code to prepare for V2
Remove unused code:
- `Files.isRegularFile`
- `Iterable<SecureHash>.summaryHash()`
- `JarExtractorInputStream`
- `parseInt`
- `resolveDependencies`
- `UncloseableOutputStream`
- `ZipExtractorInputStream`
- Package export for `net.corda.packaging.util`

Extract `ManifestCordappInfo`, `TeeInputStream` and `UncloseableInputStream` into their own files.

Remove dependency on Dev Preview 1 from test workflow cpk.
