+++ gradle.properties
@@ -15,8 +15,6 @@ kotlin.stdlib.default.dependency = false
 
 # at the moment, required for packaging to build test cpks for unit tests
 gradlePluginsVersion = 6.0.0-BETA14
-cordaDevPreviewVersion = 5.0.0-DevPreview-RC06
-
 
 # License
 ## TODO: Change to correct name and URL
