--- packaging/src/main/java/net/corda/packaging/util/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.packaging.util;
-
-import org.osgi.annotation.bundle.Export;
